requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --extra=dev --output-file=requirements.txt pyproject.toml
#
anyio==4.1.0
    # via httpx
black==23.11.0
    # via tagz (pyproject.toml)
certifi==2023.11.17
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via black
h11==0.14.0
    # via httpcore
httpcore==1.0.2
    # via httpx
httpx==0.25.2
    # via notion-client
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via tagz (pyproject.toml)
mypy-extensions==1.0.0
    # via black
notion-client==2.1.0
    # via tagz (pyproject.toml)
packaging==23.2
    # via
    #   black
    #   pytest
pathspec==0.11.2
    # via black
platformdirs==4.1.0
    # via black
pluggy==1.3.0
    # via pytest
pytest==7.4.3
    # via tagz (pyproject.toml)
python-dotenv==1.0.0
    # via tagz (pyproject.toml)
requests==2.31.0
    # via tagz (pyproject.toml)
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
urllib3==2.1.0
    # via requests
